requirements.txt
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
Django==4.2.7
django-cors-headers==4.3.1
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kafka-python==2.0.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
referencing==0.32.0
rpds-py==0.15.2
sqlparse==0.4.4
